refactored to a class
